+++ jest.config.js
@@ -1,7 +1,0 @@
-const config = {
-  preset: 'ts-jest',
-  collectCoverageFrom: ['src/*.ts'],
-  globalSetup: './global-setup.js',
-};
-
-module.exports = config;